// store/types/businessInternalUserType.ts
export interface BusinessInternalUserContact {
    key: number;
    id: string;
    first_name_th: string;
    last_name_th: string;
    // first_name_eng: string;
    // last_name_eng: string;
    id_card: string;
    email: string;
    has_ca: boolean;
    role_list: string[];
    full_name: string;

}